refactored to set up for routing
