+++ src/App.css
@@ -281,7 +281,7 @@ header {
     background-repeat: no-repeat;
     background-size: cover;
     background-position: bottom;
-    min-height: 30vh;
+    min-height: 50vh;
 }
 
 .headerContainer {
@@ -290,7 +290,7 @@ header {
     flex-direction: row;
     align-items: center;
     justify-content: center;
-    padding: 3rem 0;
+    padding: 1.5rem 0 3rem 0;
 }
 
 .headerTextContainer {
@@ -300,19 +300,29 @@ header {
     align-items: center;
     justify-content: center;
     padding-bottom: 1rem;
+    flex: 1;
 }
 
 .imgContainer {
     position: relative;
+    flex: 1;
 }
 
 @media screen and (max-width: 930px) {
-    .headerImg {
-        display: none;
+    .headerContainer {
+        flex-direction: column;
+        gap: 1rem;
     }
 }
 
 /* DASHBOARD */
+
+.dashboard {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding-top: 2rem;
+}
 
 .addTripBtn {
     font-size: 0.9rem;
@@ -338,6 +348,13 @@ header {
     box-shadow: 28px 28px 50px 0 rgba(5, 15, 31, 0.16), -23px -23px 45px 0 #FFFFFF;
     /* transition: 0.3s ease-in-out;
     opacity: 0.85; */
+}
+
+@media screen and (max-width: 930px) {
+    .card{
+        width: 330px;
+        height: 500px;
+    }
 }
 
 /* .card:hover {
