+++ src/pages/Login.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Login() {
+  return (
+    <div>Login Page</div>
+  )
+}
+
+export default Login;
