+++ src/pages/Main.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Main() {
+  return (
+    <div>Main Page</div>
+  )
+}
+
+export default Main;
