+++ src/pages/TripDetails.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function TripDetails() {
+  return (
+    <div>TripDetails</div>
+  )
+}
+
+export default TripDetails;
